Group7/requirement.txt
tflite-runtime
opencv-python
numpy
scipy
pillow
imutils
face_recognition
pygame
glob
tensorflow
opencv-python
gpiozero
Pillow

